+++ packages/zowe-explorer/src/trees/dataset/DatasetTableView.ts
@@ -228,7 +228,10 @@ export function buildMemberInfo(member: any, parentUri: string): IDataSetInfo {
  * API-based data source that directly queries the MVS API with a pattern
  */
 export class PatternDataSource implements IDataSetSource {
-    public constructor(public profile: imperative.IProfileLoaded, private pattern: string) {
+    public constructor(
+        public profile: imperative.IProfileLoaded,
+        private pattern: string
+    ) {
         this.pattern = this.pattern.toLocaleUpperCase();
     }
 
@@ -433,7 +436,9 @@ export class DatasetTableView {
             headerName: l10n.t("Creation Date"),
             useDateComparison: true,
             valueFormatter: (params: { value: string }): string => {
-                if (!params.value) {return "";}
+                if (!params.value) {
+                    return "";
+                }
                 return new Date(params.value).toLocaleDateString(this.userLocale);
             },
         },
@@ -442,7 +447,9 @@ export class DatasetTableView {
             headerName: l10n.t("Modified Date"),
             useDateComparison: true,
             valueFormatter: (params: { value: string }): string => {
-                if (!params.value) {return "";}
+                if (!params.value) {
+                    return "";
+                }
                 return new Date(params.value).toLocaleString(this.userLocale);
             },
         },