+++ packages/zowex-for-zowe-explorer/src/index.ts
@@ -1,0 +1,7 @@
+export * from "./api";
+export * from "./ConfigUtils";
+export * from "./NativeSshHelper";
+export * from "./SshClientCache";
+export * from "./SshErrorCorrelations";
+export * from "./Utilities";
+export * from "./VscePromptApi";